+++ CHANGELOG.md
@@ -3,6 +3,7 @@ _See also: [the official LibGDX changelog](https://github.com/libgdx/libgdx/blob
 #### 1.9.13-SNAPSHOT
 
 - **[UPDATE]** Updated to LibGDX 1.9.13.
+- **[UPDATE]** Updated to Kotlin 1.4.21-2.
 - **[FEATURE]** (`ktx-style`) Added `Skin.register` extension method that allows to register widget styles with the DSL.
 
 #### 1.9.12-b1